03-tools-and-setup/technology-stack/03-cloud-deployment/exercise_airflow_cloud/pyproject.toml
[project]
name = "gcloud-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Sanz3291",email = "sanju329121@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "google-cloud-storage (>=3.4.1,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "google-cloud-bigquery (>=3.38.0,<4.0.0)",
    "db-dtypes (>=1.4.3,<2.0.0)",
    "apache-airflow-providers-google (>=18.1.0,<19.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
